pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "karate"
version = "0.0.1"
dependencies = [
    "jax",
    "jaxoplanet",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "matplotlib",
    "numpyro"
]